pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "git://github.com/QUARK-framework/QUARK-plugin-tsp.git"

[project]
name = "QUARK-plugin-tsp"
version = "0.1"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = { text = "Apache-2.0" }
dependencies = ["quark-framework>=0.1.3"]
